+++ CHANGELOG.md
@@ -7,8 +7,9 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
-...
-
+### Changed
+- Add handling for empty `PGPEncryptedDataList` packets.
+  [PR#30](https://github.com/greglook/clj-pgp/pull/30)
 
 ## [1.1.1] - 2023-11-30